+++ packages/widgetbook_annotation/lib/src/addon_config.dart
@@ -1,5 +1,8 @@
+import 'package:meta/meta.dart';
+
 /// A configuration entry for a WidgetbookAddon.
-class AddonConfig {
+@optionalTypeArgs
+class AddonConfig<T> {
   /// Creates a new [AddonConfig] using a [key] and a [value].
   ///
   /// Given the following Widgetbook URL:
@@ -9,10 +12,10 @@ class AddonConfig {
   ///
   /// Then this would be:
   /// ```dart
-  /// const AddonConfigEntry('text-scale', 'factor:2.0');
+  /// const AddonConfig('text-scale', 'factor:2.0');
   /// ```
   ///
-  /// You can use other pre-defined entries for first-cla=ss addons that are
+  /// You can use other pre-defined entries for first-class addons that are
   /// provided by Widgetbook.
   const AddonConfig(this.key, this.value);
 
@@ -22,15 +25,15 @@ class AddonConfig {
   /// The value of the entry must be a query string that can be parsed by the
   /// addon. The easiest way to get the value is to use the URL query string
   /// of a Widgetbook web build.
-  final String value;
+  final T value;
 
-  MapEntry<String, String> toMapEntry() {
+  MapEntry<String, T> toMapEntry() {
     return MapEntry(key, value);
   }
 }
 
 /// [AddonConfig] for the LocalizationAddon.
-class LocalizationAddonConfig extends AddonConfig {
+class LocalizationAddonConfig extends AddonConfig<String> {
   const LocalizationAddonConfig(
     String languageTag,
   ) : super(
@@ -40,7 +43,7 @@ class LocalizationAddonConfig extends AddonConfig {
 }
 
 /// [AddonConfig] for the ThemeAddon.
-class ThemeAddonConfig extends AddonConfig {
+class ThemeAddonConfig extends AddonConfig<String> {
   const ThemeAddonConfig(
     String themeName,
   ) : super(
@@ -50,7 +53,7 @@ class ThemeAddonConfig extends AddonConfig {
 }
 
 /// [AddonConfig] for the AlignmentAddon.
-class AlignmentAddonConfig extends AddonConfig {
+class AlignmentAddonConfig extends AddonConfig<String> {
   /// Can only be one of the predefined values:
   /// - 'Top Left'
   /// - 'Top Center'
@@ -85,7 +88,7 @@ class AlignmentAddonConfig extends AddonConfig {
 }
 
 /// [AddonConfig] for the TextScaleAddon.
-class TextScaleAddonConfig extends AddonConfig {
+class TextScaleAddonConfig extends AddonConfig<String> {
   const TextScaleAddonConfig(
     double factor,
   ) : super(
@@ -95,11 +98,21 @@ class TextScaleAddonConfig extends AddonConfig {
 }
 
 /// [AddonConfig] for the ZoomAddon.
-class ZoomAddonConfig extends AddonConfig {
+class ZoomAddonConfig extends AddonConfig<String> {
   const ZoomAddonConfig(
     double zoom,
   ) : super(
           'zoom',
           'value:$zoom',
+        );
+}
+
+/// [AddonConfig] for the SemanticsAddon.
+@experimental
+class SemanticsAddonConfig extends AddonConfig<String> {
+  const SemanticsAddonConfig(bool enabled)
+      : super(
+          'semantics',
+          'enabled:$enabled',
         );
 }