+++ packages/widgetbook_cli/lib/src/core/core.dart
@@ -1,5 +1,6 @@
 export 'cli_command.dart';
 export 'cli_commands_group.dart';
+export 'cli_exception.dart';
 export 'cli_runner.dart';
 export 'context.dart';
 export 'context_manager.dart';