+++ sandbox/lib/widgetbook.dart
@@ -14,16 +14,16 @@ void main() {
 @App(
   cloudAddonsConfigs: {
     'German Dark Center x2': [
+      ViewportAddonConfig(IosViewports.iPhone12),
       LocalizationAddonConfig('de'),
       ThemeAddonConfig('Dark'),
       AlignmentAddonConfig('Center'),
       TextScaleAddonConfig(2),
       ZoomAddonConfig(2),
+      SemanticsAddonConfig(true),
       AddonConfig('custom-addon', 'name:value'),
     ],
-    'English': [
-      LocalizationAddonConfig('en'),
-    ],
+    'English': [LocalizationAddonConfig('en')],
   },
 )
 class WidgetbookApp extends StatelessWidget {
@@ -32,21 +32,12 @@ class WidgetbookApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Widgetbook.material(
-      directories: [
-        ...directories,
-        ...components,
-      ],
+      directories: [...directories, ...components],
       addons: [
+        ViewportAddon(Viewports.all),
+        SemanticsAddon(),
         next.TimeDilationAddon(),
-        next.DeviceFrameAddon([
-          Devices.ios.iPhoneSE,
-          Devices.ios.iPhone12,
-          Devices.ios.iPhone13,
-        ]),
-        next.MaterialThemeAddon({
-          'Dark': Themes.dark,
-          'Light': Themes.light,
-        }),
+        next.MaterialThemeAddon({'Dark': Themes.dark, 'Light': Themes.light}),
         next.GridAddon(),
         next.AlignmentAddon(),
         next.TextScaleAddon(),
@@ -59,14 +50,13 @@ class WidgetbookApp extends StatelessWidget {
         ),
         next.BuilderAddon(
           name: 'Bounds',
-          builder: (context, child) => Container(
-            decoration: BoxDecoration(
-              border: Border.all(
-                color: Colors.white,
+          builder:
+              (context, child) => Container(
+                decoration: BoxDecoration(
+                  border: Border.all(color: Colors.white),
+                ),
+                child: child,
               ),
-            ),
-            child: child,
-          ),
         ),
       ],
     );