+++ packages/widgetbook_annotation/CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+- **FEAT**: Add `AddonConfig` for [Semantics Addon](https://docs.widgetbook.io/addons/semantics-addon). ([#1402](https://github.com/widgetbook/widgetbook/pull/1402))
+
+## 3.3.1
+
+- **REFACTOR**: Add generic parameter to `AddonConfig`; to allow supporting non-primitive addon configs (e.g. `ViewportAddonConfig`). ([#1318](https://github.com/widgetbook/widgetbook/pull/1318))
+
 ## 3.3.0
 
 - **BREAKING**: Set minimum SDK version to 3.3.0. ([#1349](https://github.com/widgetbook/widgetbook/pull/1349))